requirements.txt
# --extra-index-url https://download.pytorch.org/whl/cu118
# torch==2.0.1 # We have to place it here otherwise it installs the CPU version
# xformers==0.0.20
# Pillow>=10.0.1
# horde_model_reference~=0.2.2
# transformers==4.30.2

# gradio
# clip-anytorch
# diffusers


#pydantic==1.10.12
#hordelib==1.6.6
#unidecode
#regex
#rembg
#windows-curses; sys_platform == 'win32'
#GitPython
#omegaconf
#setuptools
#typing_extensions


#Scribe-bridge-refactor confirmed
loguru
pyyaml
requests
psutil
pynvml